moves the API logic in a common file
